# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Recipe App API",
    "version": "1.0.0",
    "description": "An API for managing recipes, ratings, and images. Users can register, log in, post recipes, rate them, and upload recipe images."
  },
  "servers": [
    { "url": "http://localhost:8080" }
  ],
...
  "paths": {
...
    "/recipe/image": {
      "post": {
        "summary": "Upload a recipe image (auth required)",
        "tags": ["Recipe Image"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "recipeName": { "type": "string" },
                  "image": { "type": "string", "format": "binary" }
                },
                "required": ["recipeName", "image"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Image uploaded successfully" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/recipe/image/{id}": {
      "get": {
        "summary": "Get recipe image by ID (public)",
        "tags": ["Recipe Image"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Image returned" },
          "404": { "description": "Not found" }
        }
      },
      "put": {
        "summary": "Edit recipe image (auth + author required)",
        "tags": ["Recipe Image"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "recipeName": { "type": "string" },
                  "image": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Image updated successfully" },
          "403": { "description": "Forbidden" }
        }
      },
      "delete": {
        "summary": "Delete recipe image (auth + author required)",
        "tags": ["Recipe Image"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Image deleted successfully" },
          "403": { "description": "Forbidden" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}